🏗️ Extract comments list logic from `Post` component into new dedicated component
In the new feature-based architecture, comments and all related components should be stored in the dedicated feature folder.
